+++ examples/simple.rs
@@ -90,7 +90,7 @@ async fn main() -> std::io::Result<()> {
 
     // Get PoW config. Should be called everytime there's a visitor for a
     // managed site(here mcaptcha_name)
-    let work_req = system.get_pow(mcaptcha_name.into()).await.unwrap();
+    let work_req = system.get_pow(mcaptcha_name.into()).await.unwrap().unwrap();
 
     // the following computation should be done on the client but for the purpose
     // of this illustration, we are going to do it on the server it self